#1692 add admin APIs to add/remove a user from teams
