+++ modules/context/api_org.go
@@ -1,0 +1,14 @@
+// Copyright 2016 The Gogs Authors. All rights reserved.
+// Use of this source code is governed by a MIT-style
+// license that can be found in the LICENSE file.
+
+package context
+
+import (
+	"github.com/gogits/gogs/models"
+)
+
+type APIOrganization struct {
+	Organization *models.User
+	Team         *models.Team
+}
